RecordStore no longer extends LocalRecordStoreStats.
It should have been modelled as a 'has a' instead of an 'is a' relation.

Apart from that, it makes the RecordStore even more complicated than it should be.
