+++ hazelcast/src/main/java/com/hazelcast/map/impl/recordstore/AbstractRecordStore.java
@@ -31,6 +31,8 @@ import com.hazelcast.map.impl.record.Record;
 import com.hazelcast.map.impl.record.RecordComparator;
 import com.hazelcast.map.impl.record.RecordFactory;
 import com.hazelcast.map.impl.record.Records;
+import com.hazelcast.monitor.LocalRecordStoreStats;
+import com.hazelcast.monitor.impl.LocalRecordStoreStatsImpl;
 import com.hazelcast.nio.serialization.Data;
 import com.hazelcast.query.impl.Indexes;
 import com.hazelcast.query.impl.QueryableEntry;
@@ -65,10 +67,7 @@ abstract class AbstractRecordStore implements RecordStore<Record> {
     protected final MapEventJournal eventJournal;
 
     protected Storage<Data, Record> storage;
-
-    private long hits;
-    private long lastAccess;
-    private long lastUpdate;
+    protected final LocalRecordStoreStatsImpl stats = new LocalRecordStoreStatsImpl();
 
     protected AbstractRecordStore(MapContainer mapContainer, int partitionId) {
         this.name = mapContainer.getName();
@@ -83,6 +82,11 @@ abstract class AbstractRecordStore implements RecordStore<Record> {
         this.mapDataStore = mapStoreContext.getMapStoreManager().getMapDataStore(name, partitionId);
         this.lockStore = createLockStore();
         this.eventJournal = mapServiceContext.getEventJournal();
+    }
+
+    @Override
+    public LocalRecordStoreStats getLocalRecordStoreStats() {
+        return stats;
     }
 
     @Override
@@ -209,49 +213,8 @@ abstract class AbstractRecordStore implements RecordStore<Record> {
         return storage;
     }
 
-    @Override
-    public long getHits() {
-        return hits;
-    }
-
-    @Override
-    public long getLastAccessTime() {
-        return lastAccess;
-    }
-
-    @Override
-    public long getLastUpdateTime() {
-        return lastUpdate;
-    }
-
-    @Override
-    public void increaseHits() {
-        this.hits++;
-    }
-
-    @Override
-    public void increaseHits(long hits) {
-        this.hits += hits;
-    }
-
-    @Override
-    public void decreaseHits(long hits) {
-        this.hits -= hits;
-    }
-
-    @Override
-    public void setLastAccessTime(long time) {
-        this.lastAccess = Math.max(this.lastAccess, time);
-    }
-
-    @Override
-    public void setLastUpdateTime(long time) {
-        this.lastUpdate = Math.max(this.lastUpdate, time);
-    }
-
-
     protected void updateStatsOnPut(boolean newRecord, long now) {
-        setLastUpdateTime(now);
+        stats.setLastUpdateTime(now);
 
         if (!newRecord) {
             updateStatsOnGet(now);
@@ -259,17 +222,11 @@ abstract class AbstractRecordStore implements RecordStore<Record> {
     }
 
     protected void updateStatsOnPut(long hits) {
-        increaseHits(hits);
+        stats.increaseHits(hits);
     }
 
     protected void updateStatsOnGet(long now) {
-        setLastAccessTime(now);
-        increaseHits();
-    }
-
-    protected void resetStats() {
-        this.hits = 0;
-        this.lastAccess = 0;
-        this.lastUpdate = 0;
+        stats.setLastAccessTime(now);
+        stats.increaseHits();
     }
 }
